cluster-health-crew/pyproject.toml
[project]
name = "cluster-health-crew"
version = "0.1.0"
description = "Kubernetes cluster health monitoring agent built with CrewAI"
readme = "README.md"
requires-python = ">=3.10, <3.14"
dependencies = [
    "kagent-crewai>=0.9.1",
    "a2a-sdk[http-server]>=0.3.23,<1.0.0",
    "kubernetes>=31.0.0",
    "uvicorn>=0.34.3",
]

[project.scripts]
cluster-health-crew = "cluster_health_crew.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel>=0.46.3"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
